% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\title{Make an animated gif of lanscape over time}
\usage{
animate(lands, stepsBy = 1, steps = NULL, fps = 6, gifName = NULL,
  occup = 0.75)
}
\arguments{
\item{lands}{the object output of the \code{\link{run_model}} function}

\item{stepsBy}{integer number, increment value from 1 to length of steps}

\item{steps}{vector, specific steps to be animated}

\item{fps}{frames per second}

\item{gifName}{string with name of the file output}

\item{occup}{numeric between 0 and 1, if \code{occup} is not \code{NULL}, the animation will include two lines to define the south range limit of boreal and the north range limit of temperate. \code{occup} define the amount of occupancy a forest state must not occupy to determine the limit of its range distribution}
}
\value{
an animated gif file
}
\description{
This function creates an animated gif with every landscape year configuration
}
\examples{
animate(lands, stepsBy = 1, fps = 5, gifName = 'myGif', occup = 0.75)
}
