% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscape.R
\name{plot_landscape}
\alias{plot_landscape}
\title{Plot landscape}
\usage{
plot_landscape(lands, step = NULL, Title = NULL, xaxis = FALSE,
  rmBorder = TRUE, rangeLimit = NULL)
}
\arguments{
\item{lands, }{the object output from either the \code{\link{run_model}} or \code{\link{create_landscape}} functions}

\item{step}{numeric, time step to be ploted. If \code{NULL} it will plot the initial landscape for a \code{create_landscape} object, and the last time step for a \code{run_model} object.}

\item{Title}{character, title of the landscape plot}

\item{xaxis}{logical, if \code{TRUE} it will add the x axis with the annual mean temperate values.}

\item{rmBorder}{logical, if \code{TRUE} the four side borders will be removed of the plot. This option is available because the model do not calculate state prevalence in the borders.}

\item{rangeLimit, }{vector, the latitudinal position of the boreal and temperate range limit in the landscape configuration. This value is obtained from the function \code{\link{range_limit}}}
}
\description{
This function plots a specific time step landscape
}
\examples{
\dontrun{
plot_landscape(lands = output, step = 100,  Title = 'land at step 100')
}
}
