% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rangeLimitMigration.R
\name{plot_rangeLimitMigration}
\alias{plot_rangeLimitMigration}
\title{Plot range limit migration}
\usage{
plot_rangeLimitMigration(lands, rangeLimitOccup = NULL)
}
\arguments{
\item{lands}{the object output of the \code{\link{run_model}} function}

\item{rangeLimitOccup}{numeric between 0 and 1 to define the mininum row occupancy of a state in the landscape. See \code{\link{run_model}} for more details}
}
\description{
This function plots the how the range limit of boreal and temperate states varies over time
}
\examples{
\dontrun{
plot_rangeLimitMigration(lands)
}
}
