% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscape.R
\name{plot_landscape}
\alias{plot_landscape}
\title{Plot landscape}
\usage{
plot_landscape(land, nRow, nCol, Title = NULL, rmBorder = TRUE,
  rangeLimit = NULL)
}
\arguments{
\item{land}{a vector with the distribution of the four forest states in the landscape. This is found in the output list from the \code{\link{run_model}} function named 'land_TX', where X is the time step}

\item{nRow}{numeric, number of rows of the landscape. Value is found in the output list from the \code{\link{run_model}} function}

\item{Title}{character, title of the landscape plot}

\item{rmBorder}{logical, if \code{TRUE} the four side borders will be removed of the plot. This option is available because the model do not calculate state prevalence in the borders.}

\item{rangeLimit, }{vector, the latitudinal position of the boreal and temperate range limit in the landscape configuration. This value is obtained from the function \code{\link{range_limit}}}

\item{nCow}{numeric, number of columns of the landscape. Value is found in the output list from the \code{\link{run_model}} function}
}
\description{
This function plots a specific time step landscape
}
\examples{
plot_landscape(land = output[[100]], nRow = output[['nRow']], nCol = output[['nCol']], Title = 'land at step 100')
}
