% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ occupancy.R
\name{plot_occupancy}
\alias{plot_occupancy}
\title{Plot landscape occupancy}
\usage{
plot_occupancy(lands, years, spar)
}
\arguments{
\item{lands}{the object output of the \code{\link{run_model}} function}

\item{years}{vector, time steps to be plotted. Remember that mfrow set must be setted before the function in case years has more than one step}

\item{spar}{numeric, value between 0 and 1 to sooth the state occupancy}
}
\description{
This function plots the distribution of state occupancy over the Latitudinal grandient of the landscape
}
\examples{
par(mfrorw = c(1, 2))
plot_occupancy(lands, years = c(1, 100), spar = 0.2)
}
